Refactor browser image decoder.
